refactor(landing-zone): move study config to dedicated repo, rename data dir
Study config (config.yaml, validation-rules.yaml) now lives exclusively
in Appsilon/mediforce-landing-zone-study-demo — the monorepo copy was
a v0.1 holdover that would drift. Sample data for demos stays in the
monorepo, renamed from studies/CDISCPILOT01/data/ to sample-data/ for
clarity.
