+++ apps/landing-zone/README.md
@@ -68,7 +68,7 @@ From the repo root:
 docker compose -f apps/landing-zone/docker-compose.sftp.yml up -d
 ```
 
-This binds `127.0.0.1:2222` to the SFTP container and mounts `studies/CDISCPILOT01/data/sftp-staging/` as the `cro` user's `/upload` dir. Credentials are `cro` / `cro` — local dev only, not a real secret.
+This binds `127.0.0.1:2222` to the SFTP container and mounts `sample-data/sftp-staging/` as the `cro` user's `/upload` dir. Credentials are `cro` / `cro` — local dev only, not a real secret.
 
 ### Drop demo files
 
@@ -78,7 +78,7 @@ python apps/landing-zone/scripts/seed_sftp.py --variant clean
 
 Available variants: `clean`, `injection`, `mess-late`, `mess-encoding`, `mess-missing-domain`, `mess-inconsistent-values`. The `mess-late` variant additionally backdates file mtimes by 14 days to simulate an overdue delivery against `contract.expectedDeliveries[].cadence`. The script clears `sftp-staging/` first so each call models a fresh delivery.
 
-The variant directories under `studies/CDISCPILOT01/data/{variant}/` are populated by the demo data prep step (separate task); the seed script fails gracefully if they are missing.
+The variant directories under `sample-data/{variant}/` are populated by the demo data prep step (separate task); the seed script fails gracefully if they are missing.
 
 ### Test the connection
